refactor: move sequence recovery helper to sequence.go
